TargetStatus を umapita_target_status.{h,cpp} に分離
